requirements.txt
llama-cpp-python>=0.3.0
huggingface-hub>=0.23.0
pydantic>=2.0.0
pytest>=8.0.0
